[mmptest] Port link-uithread-1/2 to an NUnit LinkAll test. Partial fix for #4975.
Fixes part of https://github.com/xamarin/xamarin-macios/issues/4975.
