+++ tests/linker/mac/link all/LinkAllTest.cs
@@ -1,4 +1,8 @@
 ﻿using System;
+using System.Runtime.CompilerServices;
+using System.Threading;
+
+using AppKit;
 using Foundation;
 
 using NUnit.Framework;
@@ -24,6 +28,42 @@ namespace LinkAllTests
 			Check ("UmAlQuraCalendar", false);
 			Check ("HijriCalendar", false);
 			Check ("ThaiBuddhistCalendar", false);
+		}
+
+		[Test]
+		public void EnsureUIThreadException ()
+		{
+			// works on main/ui thread
+			NSApplication.EnsureUIThread ();
+
+			ThreadPool.QueueUserWorkItem ((v) => Tester.Test ());
+			Assert.IsTrue (Tester.mre.WaitOne (TimeSpan.FromSeconds (10)), "Successful wait");
+#if DEBUG
+			var expected_ex_thrown = true;
+#else
+			var expected_ex_thrown = false;
+#endif
+			Assert.AreEqual (expected_ex_thrown, Tester.exception_thrown, "Success");
+		}
+
+
+		class Tester : NSObject {
+			public static ManualResetEvent mre = new ManualResetEvent (false);
+			public static bool exception_thrown;
+
+			[CompilerGenerated]
+			[Export ("foo")]
+			public static void Test ()
+			{
+				try {
+					exception_thrown = false;
+					NSApplication.EnsureUIThread ();
+				} catch (AppKitThreadAccessException) {
+					exception_thrown = true;
+				} finally {
+					mre.Set ();
+				}
+			}
 		}
 	}
 }
